+++ backend/src/salsa_telescope.rs
@@ -1,99 +1,19 @@
 use crate::telescope::Telescope;
+use crate::telescope_controller::TelescopeController;
 use async_trait::async_trait;
-use chrono::{DateTime, Utc};
-use common::coords::{horizontal_from_equatorial, horizontal_from_galactic};
+use chrono::Utc;
 use common::{
-    Direction, Location, Measurement, ObservedSpectra, ReceiverConfiguration, ReceiverError,
-    TelescopeError, TelescopeInfo, TelescopeStatus, TelescopeTarget,
+    Direction, Measurement, ObservedSpectra, ReceiverConfiguration, ReceiverError, TelescopeError,
+    TelescopeInfo, TelescopeTarget,
 };
-use hex_literal::hex;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 use tokio_util::sync::CancellationToken;
 
-use std::io::{Read, Write};
-use std::net::{SocketAddr, TcpStream};
-use std::str::FromStr;
 use std::time::Duration;
 
 use rustfft::{num_complex::Complex, FftPlanner};
 use uhd::{self, StreamCommand, StreamCommandType, StreamTime, TuneRequest, Usrp};
-
-pub const LOWEST_ALLOWED_ALTITUDE: f64 = 5.0f64 / 180.0f64 * std::f64::consts::PI;
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-enum TelescopeCommand {
-    Stop,
-    Restart,
-    GetDirection,
-    SetDirection(Direction),
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-enum TelescopeResponse {
-    Ack,
-    CurrentDirection(Direction),
-}
-
-fn parse_ack_response(
-    bytes: &[u8],
-    command_name: &str,
-) -> Result<TelescopeResponse, TelescopeError> {
-    if bytes.len() == 12 && bytes[0] == 0x57 && bytes[11] == 0x20 {
-        Ok(TelescopeResponse::Ack)
-    } else {
-        Err(TelescopeError::TelescopeIOError(format!(
-            "Unexpected response to {} command: {:?}",
-            command_name, bytes,
-        )))
-    }
-}
-
-fn parse_direction_response(
-    bytes: &[u8],
-    command_name: &str,
-) -> Result<TelescopeResponse, TelescopeError> {
-    if bytes.len() == 12 && bytes[0] == 0x58 && bytes[11] == 0x20 {
-        let azimuth = rot2prog_bytes_to_angle(&bytes[1..=5]);
-        let altitude = rot2prog_bytes_to_angle(&bytes[6..=10]);
-        Ok(TelescopeResponse::CurrentDirection(Direction {
-            azimuth,
-            altitude,
-        }))
-    } else {
-        Err(TelescopeError::TelescopeIOError(format!(
-            "Unexpected response to {} command: {:?}",
-            command_name, bytes,
-        )))
-    }
-}
-
-impl TelescopeCommand {
-    fn to_bytes(&self) -> Vec<u8> {
-        match self {
-            TelescopeCommand::Stop => hex!("57000000000000000000000F20").into(),
-            TelescopeCommand::Restart => hex!("57EFBEADDE000000000000EE20").into(),
-            TelescopeCommand::GetDirection => hex!("57000000000000000000006F20").into(),
-            TelescopeCommand::SetDirection(direction) => {
-                let mut bytes = Vec::with_capacity(13);
-                bytes.extend(hex!("57"));
-                bytes.extend(rot2prog_angle_to_bytes(direction.azimuth).as_slice());
-                bytes.extend(rot2prog_angle_to_bytes(direction.altitude).as_slice());
-                bytes.extend(hex!("5F20"));
-                bytes
-            }
-        }
-    }
-
-    fn parse_response(&self, bytes: &[u8]) -> Result<TelescopeResponse, TelescopeError> {
-        match self {
-            TelescopeCommand::Stop => parse_ack_response(bytes, "stop"),
-            TelescopeCommand::Restart => parse_ack_response(bytes, "restart"),
-            TelescopeCommand::GetDirection => parse_direction_response(bytes, "get direction"),
-            TelescopeCommand::SetDirection(_) => parse_direction_response(bytes, "set direction"),
-        }
-    }
-}
 
 pub struct ActiveIntegration {
     cancellation_token: CancellationToken,
@@ -102,19 +22,9 @@ pub struct ActiveIntegration {
 
 pub struct SalsaTelescope {
     name: String,
-    controller_address: String,
     receiver_address: String,
-    timeout: Duration,
-    target: TelescopeTarget,
-    commanded_horizontal: Option<Direction>,
-    current_direction: Option<Direction>,
-    location: Location,
-    most_recent_error: Option<TelescopeError>,
+    controller: TelescopeController,
     receiver_configuration: ReceiverConfiguration,
-    connection_established: bool,
-    lowest_allowed_altitude: f64,
-    telescope_restart_request_at: Option<DateTime<Utc>>,
-    wait_time_after_restart: chrono::Duration,
     measurements: Arc<Mutex<Vec<Measurement>>>,
     active_integration: Option<ActiveIntegration>,
 }
@@ -126,60 +36,12 @@ pub fn create(
 ) -> SalsaTelescope {
     SalsaTelescope {
         name,
-        controller_address,
         receiver_address,
-        timeout: Duration::from_secs(1),
-        target: TelescopeTarget::Stopped,
-        commanded_horizontal: None,
-        current_direction: None,
-        location: Location {
-            longitude: 0.20802143022, //(11.0+55.0/60.0+7.5/3600.0) * PI / 180.0. Sign positive, handled in gmst calc
-            latitude: 1.00170457462,  //(57.0+23.0/60.0+36.4/3600.0) * PI / 180.0
-        },
-        most_recent_error: None,
+        controller: TelescopeController::new(controller_address),
         receiver_configuration: ReceiverConfiguration { integrate: false },
-        connection_established: false,
-        lowest_allowed_altitude: LOWEST_ALLOWED_ALTITUDE,
-        telescope_restart_request_at: None,
-        wait_time_after_restart: chrono::Duration::seconds(10),
         measurements: Arc::new(Mutex::new(Vec::new())),
         active_integration: None,
     }
-}
-
-fn create_connection(telescope: &SalsaTelescope) -> Result<TcpStream, std::io::Error> {
-    let address = SocketAddr::from_str(telescope.controller_address.as_str()).unwrap();
-    let stream = TcpStream::connect_timeout(&address, telescope.timeout)?;
-    stream.set_read_timeout(Some(telescope.timeout))?;
-    stream.set_write_timeout(Some(telescope.timeout))?;
-    Ok(stream)
-}
-
-fn send_command<Stream>(
-    stream: &mut Stream,
-    command: TelescopeCommand,
-) -> Result<TelescopeResponse, TelescopeError>
-where
-    Stream: Read + Write,
-{
-    stream.write(&command.to_bytes()).unwrap();
-    let mut result = vec![0; 128];
-    let response_length = stream.read(&mut result).unwrap();
-    result.truncate(response_length);
-    command.parse_response(&result)
-}
-
-fn rot2prog_bytes_to_int(bytes: &[u8]) -> u32 {
-    bytes
-        .iter()
-        .rev()
-        .enumerate()
-        .map(|(pos, &digit)| digit as u32 * 10_u32.pow(pos as u32))
-        .sum()
-}
-
-fn rot2prog_bytes_to_angle(bytes: &[u8]) -> f64 {
-    (rot2prog_bytes_to_int(bytes) as f64 / 100.0 - 360.0).to_radians()
 }
 
 // Reading the documentation of the telescope, this should be the correct way to interpret the bytes
@@ -196,17 +58,6 @@ fn rot2prog_bytes_to_int_documented(bytes: &[u8]) -> u32 {
 #[allow(dead_code)]
 fn rot2prog_bytes_to_angle_documented(bytes: &[u8]) -> f64 {
     (rot2prog_bytes_to_int_documented(bytes) as f64 / 100.0 - 360.0).to_radians()
-}
-
-fn rot2prog_angle_to_bytes(angle: f64) -> [u8; 5] {
-    let mut bytes = [0; 5];
-    let angle = ((angle.to_degrees() + 360.0) * 100.0).round();
-    bytes[0] = (angle / 10000.0) as u8 + 0x30;
-    bytes[1] = ((angle % 10000.0) / 1000.0) as u8 + 0x30;
-    bytes[2] = ((angle % 1000.0) / 100.0) as u8 + 0x30;
-    bytes[3] = ((angle % 100.0) / 10.0) as u8 + 0x30;
-    bytes[4] = (angle % 10.0) as u8 + 0x30;
-    bytes
 }
 
 fn measure_switched(
@@ -405,25 +256,18 @@ async fn measure(
 #[async_trait]
 impl Telescope for SalsaTelescope {
     async fn get_direction(&self) -> Result<Direction, TelescopeError> {
-        let mut stream = create_connection(&self)?;
-
-        self.get_current_horizontal(&mut stream).await
+        self.controller.direction().await
     }
 
     async fn get_target(&self) -> Result<TelescopeTarget, TelescopeError> {
-        Ok(self.target)
+        self.controller.target()
     }
 
     async fn set_target(
         &mut self,
         target: TelescopeTarget,
     ) -> Result<TelescopeTarget, TelescopeError> {
-        let mut stream = create_connection(&self)?;
-
-        self.target = target;
-        self.update_direction(Utc::now(), &mut stream).await?;
-
-        Ok(target)
+        self.controller.set_target(target).await
     }
 
     async fn set_receiver_configuration(
@@ -447,8 +291,8 @@ impl Telescope for SalsaTelescope {
                 })
             };
             self.active_integration = Some(ActiveIntegration {
-                cancellation_token: cancellation_token,
-                measurement_task: measurement_task,
+                cancellation_token,
+                measurement_task,
             });
         } else if !receiver_configuration.integrate && self.receiver_configuration.integrate {
             log::info!("Stopping integration");
@@ -461,22 +305,7 @@ impl Telescope for SalsaTelescope {
     }
 
     async fn get_info(&self) -> Result<TelescopeInfo, TelescopeError> {
-        let current_horizontal = match self.current_direction {
-            Some(current_horizontal) => current_horizontal,
-            None => return Err(TelescopeError::TelescopeNotConnected),
-        };
-
-        let status = match self.commanded_horizontal {
-            Some(commanded_direction) => {
-                // Check if more than 2 tolerances off, if so we are not tracking anymore
-                if directions_are_close(commanded_direction, current_horizontal, 2.0) {
-                    TelescopeStatus::Tracking
-                } else {
-                    TelescopeStatus::Slewing
-                }
-            }
-            None => TelescopeStatus::Idle,
-        };
+        let controller_info = self.controller.info()?;
 
         let latest_observation = {
             let measurements = self.measurements.lock().await;
@@ -496,36 +325,17 @@ impl Telescope for SalsaTelescope {
 
         Ok(TelescopeInfo {
             id: self.name.clone(),
-            status,
-            current_horizontal,
-            commanded_horizontal: self.commanded_horizontal,
-            current_target: self.target,
-            most_recent_error: self.most_recent_error.clone(),
+            status: controller_info.status,
+            current_horizontal: controller_info.current_horizontal,
+            commanded_horizontal: controller_info.commanded_horizontal,
+            current_target: controller_info.target,
+            most_recent_error: controller_info.most_recent_error,
             measurement_in_progress: self.active_integration.is_some(),
-            latest_observation: latest_observation,
+            latest_observation,
         })
     }
 
     async fn update(&mut self, _delta_time: Duration) -> Result<(), TelescopeError> {
-        let now = Utc::now();
-        if let Some(telescope_restart_request_at) = self.telescope_restart_request_at {
-            if now - telescope_restart_request_at < self.wait_time_after_restart {
-                return Ok(());
-            } else {
-                self.telescope_restart_request_at = None;
-            }
-        }
-
-        let stream = create_connection(&self);
-        let mut stream = match stream {
-            Ok(stream) => stream,
-            Err(telescope_error) => {
-                let error: TelescopeError = telescope_error.into();
-                self.most_recent_error = Some(error.clone());
-                return Err(error);
-            }
-        };
-
         if let Some(active_integration) = self.active_integration.take() {
             if active_integration.measurement_task.is_finished() {
                 if let Err(error) = active_integration.measurement_task.await {
@@ -535,104 +345,12 @@ impl Telescope for SalsaTelescope {
                 self.active_integration = Some(active_integration);
             }
         }
-
-        self.update_direction(now, &mut stream).await?;
         Ok(())
     }
 
     async fn restart(&mut self) -> Result<(), TelescopeError> {
-        let mut stream = create_connection(&self)?;
-        send_command(&mut stream, TelescopeCommand::Restart)?;
-        self.most_recent_error = None;
-        self.connection_established = false;
-        self.telescope_restart_request_at = Some(Utc::now());
+        self.controller.restart();
         Ok(())
-    }
-}
-
-fn directions_are_close(a: Direction, b: Direction, tol: f64) -> bool {
-    // The salsa telescope works with a precision of 0.1 degrees
-    // We want to send new commands whenever we exceed this tolerance
-    // but to report tracking status we allow more, so that we do not flip
-    // status between tracking/slewing (e.g. due to control unit rounding errors)
-    // Therefore we have the "tol" multiplier here, which scales the allowed error.
-    let epsilon = tol * 0.1_f64.to_radians();
-    (a.azimuth - b.azimuth).abs() < epsilon && (a.altitude - b.altitude).abs() < epsilon
-}
-
-impl SalsaTelescope {
-    fn calculate_target_horizontal(&self, when: DateTime<Utc>) -> Option<Direction> {
-        match self.target {
-            TelescopeTarget::Equatorial { ra, dec } => {
-                Some(horizontal_from_equatorial(self.location, when, ra, dec))
-            }
-            TelescopeTarget::Galactic { l, b } => {
-                Some(horizontal_from_galactic(self.location, when, l, b))
-            }
-            TelescopeTarget::Stopped => None,
-            TelescopeTarget::Parked => None,
-        }
-    }
-
-    async fn get_current_horizontal<Stream>(
-        &self,
-        stream: &mut Stream,
-    ) -> Result<Direction, TelescopeError>
-    where
-        Stream: Read + Write,
-    {
-        match send_command(stream, TelescopeCommand::GetDirection)? {
-            TelescopeResponse::CurrentDirection(direction) => Ok(direction),
-            _ => Err(TelescopeError::TelescopeIOError(
-                "Telescope did not respond with current direction".to_string(),
-            )),
-        }
-    }
-
-    async fn update_direction<Stream>(
-        &mut self,
-        when: DateTime<Utc>,
-        stream: &mut Stream,
-    ) -> Result<(), TelescopeError>
-    where
-        Stream: Read + Write,
-    {
-        let target_horizontal = self.calculate_target_horizontal(when);
-        let current_horizontal = self.get_current_horizontal(stream).await?;
-        self.current_direction = Some(current_horizontal);
-
-        if !self.connection_established {
-            send_command(stream, TelescopeCommand::Stop)?;
-            self.commanded_horizontal = None;
-            self.connection_established = true;
-            return Ok(());
-        }
-
-        match target_horizontal {
-            Some(target_horizontal) => {
-                if target_horizontal.altitude < self.lowest_allowed_altitude {
-                    self.most_recent_error = Some(TelescopeError::TargetBelowHorizon);
-                    self.commanded_horizontal = None;
-                    return Err(TelescopeError::TargetBelowHorizon);
-                }
-
-                self.commanded_horizontal = Some(target_horizontal);
-
-                // Check if more than 1 tolerance off, if so we need to send track command
-                if !directions_are_close(target_horizontal, current_horizontal, 1.0) {
-                    send_command(stream, TelescopeCommand::SetDirection(target_horizontal))?;
-                }
-
-                Ok(())
-            }
-            None => {
-                if self.commanded_horizontal.is_some() {
-                    send_command(stream, TelescopeCommand::Stop)?;
-                    self.commanded_horizontal = None;
-                }
-                Ok(())
-            }
-        }
     }
 }
